feat: add new hooks option
